fix: broadcast guardrails — message length, past schedule, half-second delay
- Cap broadcast message length at 4096 (Telegram's limit) in the API, so
  a template/duplicate/API edit can't smuggle a too-long message that
  Telegram would reject for every recipient (M14).
- Reject scheduling a broadcast in the past — it would fire instantly but
  be labelled "scheduled" (m24).
- Accept a fractional delay-between-messages so the 0.5s slider step no
  longer 422s (m12).

+++ apps/api/src/routes/settings.ts
@@ -32,7 +32,8 @@ const messengerParamSchema = z.object({
 
 const updateAntibanBodySchema = z.object({
   messagesPerBatch: z.number().int().min(1).max(1000),
-  delayBetweenMessages: z.number().int().min(0).max(3600),
+  // Float: the UI slider steps by 0.5s, so an int() schema 422'd on half-seconds.
+  delayBetweenMessages: z.number().min(0).max(3600),
   delayBetweenBatches: z.number().int().min(0).max(86400),
   maxMessagesPerHour: z.number().int().min(1).max(10000),
   maxMessagesPerDay: z.number().int().min(1).max(100000),
@@ -44,7 +45,7 @@ const updateAntibanBodySchema = z.object({
 const riskScoreQuerySchema = z.object({
   messenger: messengerEnum,
   messagesPerBatch: z.coerce.number().int().min(1),
-  delayBetweenMessages: z.coerce.number().int().min(0),
+  delayBetweenMessages: z.coerce.number().min(0),
   delayBetweenBatches: z.coerce.number().int().min(0),
   maxMessagesPerHour: z.coerce.number().int().min(1),
   maxMessagesPerDay: z.coerce.number().int().min(1),
